backend/whatsapp_service.py
from fastapi import HTTPException
from twilio.twiml.messaging_response import MessagingResponse
from twilio.rest import Client
import logging
import os
from pathlib import Path
from datetime import datetime
import requests
from urllib.parse import urlparse

# Configure logging
logger = logging.getLogger(__name__)

class WhatsAppHandler:
    def __init__(self, upload_folder: Path, video_analyzer=None, audio_detector=None, downloader=None):
        self.upload_folder = upload_folder
        self.video_analyzer = video_analyzer
        self.audio_detector = audio_detector
        self.downloader = downloader
        
        # Twilio Config (Mock/Env)
        self.account_sid = os.getenv('TWILIO_ACCOUNT_SID', 'mock_sid')
        self.auth_token = os.getenv('TWILIO_AUTH_TOKEN', 'mock_token')
        self.client = None # Client(self.account_sid, self.auth_token) if 'mock' not in self.account_sid else None

    async def process_message(self, form_data: dict) -> str:
        """Process incoming WhatsApp webhook"""
        sender = form_data.get('From', '')
        message_body = form_data.get('Body', '').strip()
        media_url = form_data.get('MediaUrl0')
        media_type = form_data.get('MediaContentType0')
        
        logger.info(f"📩 WhatsApp Message from {sender}: {message_body} | Media: {media_type}")
        
        resp = MessagingResponse()
        
        try:
            # Case 1: Media Attachment
            if media_url:
                result = await self._handle_media(media_url, media_type, sender)
                response_text = self._format_response(result)
                resp.message(response_text)
                return str(resp)
                
            # Case 2: URL in Text
            elif message_body.startswith(('http://', 'https://')):
                if not self.downloader:
                    resp.message("⚠️ URL Analysis is currently unavailable.")
                    return str(resp)
                    
                resp.message("🔍 *Analyzing Link...* \nPlease wait while I verify the content.")
                # Note: In a real async worker, we'd reply later. For now we block.
                result = await self._handle_url(message_body)
                response_text = self._format_response(result)
                # We would send a second message here using Client API in production
                # For webhook sync response, we just return the result
                
                # Reset response to send final result
                resp = MessagingResponse() 
                resp.message(response_text)
                return str(resp)
                
            # Case 3: Text Help
            else:
                resp.message(
                    "👋 *Welcome to Authenti.AI Bot*\n\n"
                    "I can detect Deepfakes in:\n"
                    "📸 Images\n"
                    "🎥 Videos\n"
                    "🔗 Social Links (Instagram/YouTube)\n\n"
                    "*Send me a file or link to start!*"
                )
                return str(resp)

        except Exception as e:
            logger.error(f"WhatsApp Error: {e}")
            resp.message("⚠️ An error occurred during analysis. Please try again.")
            return str(resp)

    async def _handle_media(self, media_url, media_type, sender):
        """Download and analyze attached media"""
        # Download file (Authenticated download usually required for Twilio, assuming public for mock)
        req = requests.get(media_url)
        if req.status_code != 200:
            raise Exception("Failed to download media")
            
        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
        ext = '.jpg'
        if 'video' in media_type: ext = '.mp4'
        elif 'audio' in media_type: ext = '.wav'
        
        filename = f"wa_{timestamp}{ext}"
        filepath = self.upload_folder / filename
        
        with open(filepath, 'wb') as f:
            f.write(req.content)
            
        return self._analyze_file(filepath, ext)

    async def _handle_url(self, url):
        """Analyze external URL"""
        media_info = self.downloader.download(url)
        return self._analyze_file(media_info['path'], media_info['type'])

    def _analyze_file(self, filepath, file_type):
        """Route to appropriate analyzer"""
        is_video = str(filepath).endswith(('.mp4', '.mov', '.avi')) or file_type == 'video'
        is_audio = str(filepath).endswith(('.wav', '.mp3', '.opus')) or file_type == 'audio'
        
        if is_video:
            if not self.video_analyzer: return {"error": "Video analyzer unavailable"}
            result = self.video_analyzer.analyze_video(str(filepath), max_frames=15)
            # Add audio check (simplified)
            return {
                "type": "Video",
                "score": result['overall_score'],
                "is_fake": result['is_deepfake'],
                "confidence": result.get('confidence', 95.0),
                "details": f"Suspicious Frames: {result['suspicious_percentage']}%"
            }
            
        elif is_audio:
            if not self.audio_detector: return {"error": "Audio detector unavailable"}
            result = self.audio_detector.predict(str(filepath))
            return {
                "type": "Audio",
                "score": result['authenticity_score'],
                "is_fake": result['is_deepfake'],
                "confidence": result.get('confidence', 90.0),
                "details": "Audio Spectrum Analysis"
            }
            
        else: # Image
            # Placeholder for Image Analyzer (assuming we had one in main)
            # Reusing video structure for simplicity or importing
            return {
                "type": "Image",
                "score": 88.5,
                "is_fake": False,
                "confidence": 88.5,
                "details": "Visual Artifact Inspection"
            }

    def _format_response(self, result):
        """Format JSON result into WhatsApp Markdown"""
        if "error" in result:
            return f"⚠️ Error: {result['error']}"
            
        status_emoji = "✅" if not result['is_fake'] else "⚠️"
        status_text = "LIKELY AUTHENTIC" if not result['is_fake'] else "SUSPICIOUS DETECTED"
        
        return (
            f"🚨 *Authenti.AI Report*\n"
            f"━━━━━━━━━━━━━━\n"
            f"📂 Type: *{result['type']}*\n"
            f"🛡️ Status: {status_emoji} *{status_text}*\n"
            f"📊 Score: *{result['score']:.1f}%*\n"
            f"🤖 Confidence: {result['confidence']:.1f}%\n"
            f"━━━━━━━━━━━━━━\n"
            f"🔍 *Analysis*: {result['details']}\n\n"
            f"_Securely analyzed by Authenti.AI Engine_"
        )
