backend/requirements.txt
fastapi
pytz
transformers
torch
timm
Pillow
numpy
opencv-python
torchvision
facenet-pytorch
librosa
soundfile
yt-dlp
twilio
python-multipart
uvicorn